[domain] move Config into module
