+++ src/main/scala/net/kemitix/s3thorp/ActionSubmitter.scala
@@ -2,6 +2,7 @@ package net.kemitix.s3thorp
 
 import cats.effect.IO
 import net.kemitix.s3thorp.awssdk.{S3Client, UploadProgressListener}
+import net.kemitix.s3thorp.domain.Config
 
 trait ActionSubmitter
   extends S3Client
