+++ src/main/scala/net/kemitix/s3thorp/Logging.scala
@@ -1,6 +1,7 @@
 package net.kemitix.s3thorp
 
 import com.typesafe.scalalogging.LazyLogging
+import net.kemitix.s3thorp.domain.Config
 
 trait Logging extends LazyLogging {
