+++ src/main/scala/net/kemitix/s3thorp/MD5HashGenerator.scala
@@ -3,7 +3,7 @@ package net.kemitix.s3thorp
 import java.io.{File, FileInputStream}
 import java.security.MessageDigest
 
-import net.kemitix.s3thorp.domain.MD5Hash
+import net.kemitix.s3thorp.domain.{Config, MD5Hash}
 
 trait MD5HashGenerator
   extends Logging {
