+++ src/main/scala/net/kemitix/s3thorp/awssdk/S3ClientCopier.scala
@@ -2,8 +2,8 @@ package net.kemitix.s3thorp.awssdk
 
 import cats.effect.IO
 import com.github.j5ik2o.reactive.aws.s3.cats.S3CatsIOClient
-import net.kemitix.s3thorp.domain.{Bucket, MD5Hash, RemoteKey}
-import net.kemitix.s3thorp.{Config, CopyS3Action}
+import net.kemitix.s3thorp.domain.{Bucket, Config, MD5Hash, RemoteKey}
+import net.kemitix.s3thorp.CopyS3Action
 import software.amazon.awssdk.services.s3.model.CopyObjectRequest
 
 class S3ClientCopier(s3Client: S3CatsIOClient)
