+++ src/main/scala/net/kemitix/s3thorp/awssdk/S3ClientUploader.scala
@@ -2,8 +2,8 @@ package net.kemitix.s3thorp.awssdk
 
 import cats.effect.IO
 import com.amazonaws.event.{ProgressEvent, ProgressEventType, ProgressListener}
-import net.kemitix.s3thorp.domain.{Bucket, LocalFile}
-import net.kemitix.s3thorp.{Config, S3Action}
+import net.kemitix.s3thorp.domain.{Bucket, Config, LocalFile}
+import net.kemitix.s3thorp.S3Action
 
 trait S3ClientUploader {
