+++ src/test/scala/net/kemitix/s3thorp/DummyS3Client.scala
@@ -2,7 +2,7 @@ package net.kemitix.s3thorp
 
 import cats.effect.IO
 import net.kemitix.s3thorp.awssdk.{S3Client, UploadProgressListener}
-import net.kemitix.s3thorp.domain.{Bucket, LocalFile, MD5Hash, RemoteKey, S3ObjectsData}
+import net.kemitix.s3thorp.domain.{Bucket, Config, LocalFile, MD5Hash, RemoteKey, S3ObjectsData}
 
 trait DummyS3Client extends S3Client {
