+++ src/test/scala/net/kemitix/s3thorp/KeyGeneratorSuite.scala
@@ -2,7 +2,7 @@ package net.kemitix.s3thorp
 
 import java.io.File
 
-import net.kemitix.s3thorp.domain.{Bucket, RemoteKey}
+import net.kemitix.s3thorp.domain.{Bucket, Config, RemoteKey}
 import org.scalatest.FunSpec
 
 class KeyGeneratorSuite extends FunSpec {
